fix(core): fix nested generic parsing and toDouble return type consistency
- Parser.cpp: handle RSHIFT ('>>') in parseTypeName() by splitting it into
  two GREATER_THAN tokens, fixing 'Expected >' errors for nested generics
  like Map<Text, List<Int>> and List<Map<Text, Set<Int>>>
- MethodCallNode.cpp: fix Text.toDouble() and Bool.toDouble() to return
  Double instead of Float, consistent with Int/Long.toDouble()
- test_type_conversions.cpp: update all toDouble assertions and fixtures
  to expect Double (not Float); guard LongToIntConversion with O2L_HAS_INT128

+++ src/AST/MethodCallNode.cpp
@@ -1720,7 +1720,7 @@ Value MethodCallNode::evaluate(Context& context) {
                     }
 
                     double result = std::stod(trimmed);
-                    return Float(result);
+                    return Double(result);
                 } catch (const std::exception&) {
                     throw EvaluationError("Cannot convert '" + text_value + "' to Double", context);
                 }
@@ -2007,7 +2007,7 @@ Value MethodCallNode::evaluate(Context& context) {
                 if (!arg_values.empty()) {
                     throw EvaluationError("Bool.toDouble() takes no arguments", context);
                 }
-                return Float(bool_value ? 1.0 : 0.0);
+                return Double(bool_value ? 1.0 : 0.0);
             } else if (method_name_ == "toFloat") {
                 if (!arg_values.empty()) {
                     throw EvaluationError("Bool.toFloat() takes no arguments", context);
